Merge remote-tracking branch 'origin/master' into auth-fixes

+++ src/main/java/com/heroku/spacey/dao/LoginInfoDao.java
@@ -4,5 +4,5 @@ import com.heroku.spacey.entity.LoginInfo;
 
 public interface LoginInfoDao {
     LoginInfo getLoginInfoByEmail(String email);
-    long insert(LoginInfo loginInfo);
+    int insert(LoginInfo loginInfo);
 }
